requirements.txt
asgiref==3.6.0
asyncio==3.4.3
autopep8==2.0.1
boto3==1.26.59
botocore==1.29.59
Django==4.1.5
django-crispy-forms==1.14.0
django-extra-views==0.14.0
django-storages==1.13.2
Faker==16.6.1
flake8==6.0.0
jmespath==1.0.1
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.14
